Traveller Agent App/requirements.txt
google-generativeai==0.3.2
asyncio
typing
dataclasses
json
random
datetime
uuid 